requirements.txt
blessings==1.6
bpython==0.14.1
curtsies==0.1.19
dj-static==0.0.6
Django==1.7.7
djangorestframework==3.1.1
greenlet==0.4.5
Pygments==2.0.2
requests==2.6.0
six==1.9.0
static3==0.5.1
